+++ wemake_python_styleguide/logic/complexity/overuses.py
@@ -79,7 +79,7 @@ def is_primitive(node: ast.AST) -> bool:
     We do not check for strings, numbers, etc
     because they are globally ignored.
     """
-    if isinstance(node, (ast.Tuple, ast.List)):
+    if isinstance(node, ast.Tuple | ast.List):
         return not node.elts  # we do allow `[]` and `()`
     if isinstance(node, ast.Set):
         elts = node.elts
@@ -105,7 +105,7 @@ def is_unary_minus(node: ast.AST) -> bool:
     """
     if isinstance(node, ast.UnaryOp) and isinstance(node.op, ast.USub):
         # We allow variables, attributes, subscripts, and `-1`
-        if isinstance(node.operand, (ast.Constant, ast.Num)):
+        if isinstance(node.operand, ast.Constant | ast.Num):
             return node.operand.n == 1
         return True
     return False