+++ wemake_python_styleguide/logic/tree/attributes.py
@@ -6,7 +6,7 @@ from wemake_python_styleguide.types import AnyChainable, AnyVariableDef
 
 
 def _chained_item(iterator: ast.AST) -> ast.AST | None:
-    if isinstance(iterator, (ast.Attribute, ast.Subscript)):
+    if isinstance(iterator, ast.Attribute | ast.Subscript):
         return iterator.value
     if isinstance(iterator, ast.Call):
         return iterator.func