+++ wemake_python_styleguide/visitors/decorators.py
@@ -38,7 +38,7 @@ def alias(
     We can just create aliases like ``visit_Import = visit_ImportFrom``,
     but it looks verbose and ugly.
     """
-    all_names = aliases + (original,)
+    all_names = (*aliases, original)
     if len(all_names) != len(set(all_names)):
         raise ValueError('Found duplicate aliases')